Tests Completed!
All tests are completed and passing with no errors or failures! The only
methods not tested are SocialNetwork.loadFromFile() &
SocialNetwork.saveToFile() as well as the two QOL methods I added
earlier this morning which shouldn't need testing.

At this point I think we can move on to implementing GUI functionality.

Let's Wrap this up!