I'm Done
I think that's it for me guys. I got it mostly functional. Its too much
effort to refactor so much code to make everything perfect and track
down all the bugs. Let me know if you can break it. Hopefully this is
satisfactory. Still need to get a submit to canvas done.